SDK/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
base58==2.1.1
bitarray==2.6.1
blinker==1.4
cachetools==5.2.0
cairo-lang==0.10.3
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
cloud-init==20.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cytoolz==0.12.0
dbus-python==1.2.16
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
ecdsa==0.18.0
entrypoints==0.3
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
exceptiongroup==1.0.4
fastecdsa==2.2.3
filelock==3.8.2
frozendict==1.2
frozenlist==1.3.3
hexbytes==0.3.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
lark==1.1.5
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lru-dict==1.1.8
MarkupSafe==1.1.0
marshmallow==3.19.0
marshmallow-dataclass==8.5.10
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
more-itertools==4.2.0
mpmath==1.2.1
multiaddr==0.0.9
multidict==6.0.3
mypy-extensions==0.4.3
netaddr==0.8.0
netifaces==0.10.4
numpy==1.24.0
oauthlib==3.1.0
packaging==22.0
parsimonious==0.8.1
pipdeptree==2.3.3
platformdirs==2.6.0
pluggy==1.0.0
prometheus-client==0.15.0
protobuf==3.19.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycryptodome==3.16.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
pytest==7.2.0
pytest-asyncio==0.20.3
python-apt==2.0.0
python-debian===0.1.36ubuntu1
python-dotenv==1.0.0
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
rlp==2.0.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.10
sympy==1.11.1
systemd-python==234
tomli==2.0.1
toolz==0.12.0
Twisted==18.9.0
typeguard==2.13.3
typing-extensions==4.4.0
typing-inspect==0.8.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
varint==1.0.2
virtualenv==20.17.1
wadllib==1.3.3
web3==5.31.3
websockets==9.1
yarl==1.8.2
zipp==1.0.0
zope.interface==4.7.1
